+++ src/main/java/com/example/teamproject/domain/vo/ProductReplyVO.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Component;
 @Component
 @Data
 public class ProductReplyVO {
-    private int prno;
-    private int pno;
+    private Long prno;
+    private Long pno;
     private String content;
     private String name;
     private String registerDate;
