+++ src/main/java/com/example/teamproject/domain/vo/RequestReplyVO.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Component;
 @Component
 @Data
 public class RequestReplyVO {
-    private int rrno;
-    private int rno;
+    private Long rrno;
+    private Long rno;
     private String content;
     private String name;
     private String registerDate;
